requirements.txt
langchain
langchain-community

ipykernel
python-dotenv

pypdf
beautifulsoup4
arxiv
pymupdf
wikipedia

langchain-text-splitters
lxml

langchain-openai
langchain-huggingface
sentence-transformers

faiss-cpu
chromadb
langchain-chroma

streamlit

langchain-groq
langchain-core
langserve
langchain-ollama

fastapi
uvicorn
sse_starlette
pydantic==1.10.13
tqdm

arxiv
wikipedia
duckduckgo-search

mysql-connector-python
SQLAlchemy

validators
youtube-transcript-api
unstructured
pytube
numexpr
huggingface-hub
gradio
openai
langchain-nvidia-ai-endpoints